Move Synth class to lib/
