+++ lib/Synth.js
